Allow weights to be set on v3 route destinations
Weights are only valid on PATCH /v3/routes/:guid/destinations,
because when using weighted routing, weights of all destinations for a
route must add up 100. Managing that restriction through insert/delete
endpoints is impossible.

We improved error messages for malformed requests to destinations
endpoints. Now, errors that are specific to a single destination object
in the request will be reported like `Destinations[1]: must be a hash.`.

[#165842090]

Co-authored-by: Luan Santos <lsantos@pivotal.io>
Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>