Start entombing VCAP
- Create new vcap_sarcophagus directory to confine code relating to the
the "vcap framework"
- The objective is to separate code responsible for cloud controller
business logic from framework code that can be replaced with Rails
- This will let us delete the code in the sarcophagus once we are able
to remove the v2 API (and move fully to Rails)
- The distinction between "business logic" code and "framework" code is
fuzzy at best, so this will likely take multiple iterations to get
right. Hopefully this will help us better track calls that cross the
boundary between the sarcophagus and the rest of the codebase.

Authored-by: Greg Cobb <gcobb@pivotal.io>