Allow UAA clients to be created with POST /v3/users endpoint
For UAA clients, the `presentation_name` field on the user resource is
the user guid, rather than the username. If a user is created through
this endpoint with a guid that does not match one in UAA's database, it
will be treated like a client (its `presentation_name` will be its
guid).

[Finishes #164231883]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Mona Mohebbi <mmohebbi@pivotal.io>